+++ src/components/Rating/Rating.styled.jsx
@@ -2,7 +2,7 @@ import styled from 'styled-components'
 import star from '../../assets/star.svg'
 import selectedStar from '../../assets/selected-star.svg'
 
-const StarBtn = styled.img`
+export const StarBtn = styled.img`
   display: inline-block;
   width: 32px;
   height: auto;
@@ -16,12 +16,10 @@ const StarBtn = styled.img`
   }
 `
 
-const SelectedStarBtn = styled.img`
+export const SelectedStarBtn = styled.img`
   display: inline-block;
   width: 32px;
   height: auto;
   margin: 5px;
   content: url(${selectedStar});
 `
-
-export { StarBtn, SelectedStarBtn }
